将 Tile 从 Sprite 提取出来

+++ src/Clem/Renderer/Color.cpp
@@ -3,13 +3,13 @@
 
 #include "Color.h"
 
-Color::Color(short color)
+Color::Color(uint16_t color)
 {
 	fore = color & 0x0F;
 	back = color & 0xF0;
 }
 
-Color::Color(short f, short b)
+Color::Color(uint8_t f, uint8_t b)
 		: fore(f), back(b)
 {
 }
